Update plone.dublincore behavior fields even if the object doesn't have this behavior
Plone fills in and uses the plone.dublincore behavior fields, such as,
effective, expires, creators..., even if the object doesn't have this
behavior. So content exports must bring these fields even if the object
doesn't have the behavior. These changes import these fields in object
that doesn't have this behavior, if they are in json.